+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+
+* **Breaking**: [398](https://github.com/SimformSolutionsPvtLtd/chatview/pull/398)
+  Implement logic to manage voice message playback during recording and manage currently playing voice messages
+
 ## [3.0.0]
 
 * **Fix**: [415](https://github.com/SimformSolutionsPvtLtd/chatview/pull/415)